updates playwright config
for added monitoring, update the playwright config to make use of the html
reporter, and the github reporter when in CI (or will fallback to the
list reporter when not in CI).

also updates the content security policy to allow for any port to be
used on localhost for both app.html and playwright/index.html.