issue/79 - sparks animation optimization
The wrong property was being used when attempting to remove the spark
from the parent object, so the sparks weren't actually being removed.
This was causing too many sparks to be added, resulting in frame rates
tanking as the number of sparks on the object increased.

This update fixes the incorrect property from being used, so the expired
sparks are actually removed. This update fixes the frame rate decline
issue.